feat: Correctly position the video according to the zoom position and matrix size
