+++ macos/Flutter/ephemeral/Flutter-Generated.xcconfig
@@ -3,8 +3,8 @@ FLUTTER_ROOT=C:\Users\bruno\Documents\flutter\flutter
 FLUTTER_APPLICATION_PATH=C:\Users\bruno\Documents\flutter\paid_projects\unity
 COCOAPODS_PARALLEL_CODE_SIGN=true
 FLUTTER_BUILD_DIR=build
-FLUTTER_BUILD_NAME=3.0.014
-FLUTTER_BUILD_NUMBER=3.0.014
+FLUTTER_BUILD_NAME=3.0.015
+FLUTTER_BUILD_NUMBER=3.0.015
 DART_OBFUSCATION=false
 TRACK_WIDGET_CREATION=true
 TREE_SHAKE_ICONS=false
