+++ macos/Flutter/ephemeral/flutter_export_environment.sh
@@ -4,8 +4,8 @@ export "FLUTTER_ROOT=C:\Users\bruno\Documents\flutter\flutter"
 export "FLUTTER_APPLICATION_PATH=C:\Users\bruno\Documents\flutter\paid_projects\unity"
 export "COCOAPODS_PARALLEL_CODE_SIGN=true"
 export "FLUTTER_BUILD_DIR=build"
-export "FLUTTER_BUILD_NAME=3.0.014"
-export "FLUTTER_BUILD_NUMBER=3.0.014"
+export "FLUTTER_BUILD_NAME=3.0.015"
+export "FLUTTER_BUILD_NUMBER=3.0.015"
 export "DART_OBFUSCATION=false"
 export "TRACK_WIDGET_CREATION=true"
 export "TREE_SHAKE_ICONS=false"
